requirements.txt
google-cloud-storage
flask
functions_framework